Restructure

+++ Trees/Delete Tree Function/tree_use.cpp
@@ -39,6 +39,9 @@ TreeNode<int>* takeInputLevelWise() {
    Deletion is done in a post order traversal way.
  */
 void deleteTree(TreeNode<int>* root){
+	if(root == NULL){
+		return;
+	}
 	for(int i=0 ; i < root->children.size() ; i++){
 		deleteTree(root->children[i]);
 	}
